refactor(docs): simplify script decoupling
