+++ docs/website/package.json
@@ -3,7 +3,7 @@
     "examples": "docusaurus-examples",
     "start": "docusaurus start",
     "build": "update_plugins_docs && docusaurus build && cp _redirects build",
-    "update_plugins_docs": "./node_modules/.bin/ts-node ./scripts/update-plugins-docs.ts; npm run prettify_plugins_docs;",
+    "update_plugins_docs": "./node_modules/.bin/ts-node ./scripts/index.ts; npm run prettify_plugins_docs;",
     "prettify_plugins_docs": "../../node_modules/.bin/prettier --write --config ../../.prettierrc ./sidebars.json ./i18n/en.json ../docs/plugins",
     "swizzle": "docusaurus swizzle",
     "deploy": "docusaurus deploy"
