+++ docs/website/scripts/file-system.ts
@@ -1,13 +1,13 @@
 import { readFileSync, readdirSync, writeFileSync } from 'fs'
-import { ENCODINGS } from '../constants'
+
+const UTF8: BufferEncoding = 'utf-8'
 
 export const readJSON = (filePath: string): void =>
-  JSON.parse(readFileSync(filePath, ENCODINGS.UTF8 as BufferEncoding))
+  JSON.parse(readFileSync(filePath, UTF8))
 
 export const writeJSON = (filePath: string, data): void =>
   writeFileSync(filePath, JSON.stringify(data, null, 2))
 
 export const readDir = readdirSync
-export const readFile = (path: string): string =>
-  readFileSync(path, ENCODINGS.UTF8 as BufferEncoding)
+export const readFile = (path: string): string => readFileSync(path, UTF8)
 export const writeFile = writeFileSync
