--- docs/website/scripts/types.ts
@@ -1,4 +1,0 @@
-export interface PluginInfo {
-  id: string
-  title: string
-}
