--- docs/website/scripts/util/formatting.ts
@@ -1,22 +1,0 @@
-export const capitalize = string =>
-  string.charAt(0).toUpperCase() + string.slice(1)
-
-export const generateHeader = (
-  packageName: string,
-  id: string,
-  title: string
-) => {
-  const header = `---
-title: ${title}
-id: ${id}
----
-
----
-
-For more information, refer to **[GitHub](https://github.com/hubtype/botonic/tree/master/packages/${packageName})**.
-
----
-
-`
-  return header
-}
